Refactor ContractDetails to utilize EditableContractName component
- Extracted the editable contract name functionality into a new EditableContractName component for improved code organization and reusability.
- Updated ContractDetails to integrate the new component, enhancing maintainability and readability.
- Removed redundant code related to contract name editing from ContractDetails, promoting a cleaner component structure.
